refactor: move `AddImageLinksView` to `views/source_bulk_actions`
Move and refactor related files.
