+++ django/cantusdb_project/main_app/templates/source_bulk_actions/add_image_links.html
@@ -5,7 +5,7 @@
 {% endblock %}
 {% block scripts %}
     <script src="{% static 'js/source_add_image_links.js' %}"></script>
-    <link rel="stylesheet" href="{% static 'css/source_add_image_links.css' %}" />
+    <link rel="stylesheet" href="{% static 'css/source_bulk_add.css' %}" />
 {% endblock %}
 {% block content %}
     <div class="container bg-white rounded">
